fix(backend): resolve MySQL FK type mismatches across 9 models
- Replace SQLITE_BIGINT/db.BigInteger with db.Integer in analysis_run,
  article_embedding, article_snapshot, document_features,
  event_aggregation, hotspot, processing_log, sentiment, event
- Delete mock/admin.ts (real backend now handles /api/admin/*)
- Backend verified: 57 API routes, login/admin/events all 200 OK
  with MySQL database

+++ backend/app/models/event.py
@@ -13,7 +13,7 @@ class Event(db.Model):
     key_figures = db.Column(db.String(500))
     heat_index = db.Column(db.Float, default=0)
     current_heat_snapshot_id = db.Column(
-        db.BigInteger,
+        db.Integer,
         db.ForeignKey(
             "event_heat_snapshot.id",
             name="fk_event_current_heat_snapshot",
@@ -35,7 +35,7 @@ class Event(db.Model):
     sentiment_positive = db.Column(db.Float, default=0)
     sentiment_negative = db.Column(db.Float, default=0)
     sentiment_neutral = db.Column(db.Float, default=0)
-    current_sentiment_snapshot_id = db.Column(db.BigInteger, nullable=True, index=True)
+    current_sentiment_snapshot_id = db.Column(db.Integer, nullable=True, index=True)
     sentiment_score = db.Column(db.Float, default=0)
     sentiment_updated_at = db.Column(db.DateTime)
     lifecycle_stage = db.Column(db.String(20), default="潜伏期")
